requirements.txt
# FastAPI
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# Database
sqlalchemy
asyncpg
alembic

# Authentication
python-jose[cryptography]
passlib[bcrypt]
bcrypt<5.0  # Pin to 4.x for compatibility with passlib 1.7.4
python-multipart

# LangChain (for LLM integration)
langchain
langchain-openai
openai
httpx

# Environment
python-dotenv
